String serialization figured out

+++ source/ncconvert/converters/GeometryConverter.cpp
@@ -154,7 +154,7 @@ auto GetBoneWeights(const aiMesh* mesh) -> std::unordered_map<uint32_t, nc::asse
     return out;
 }
 
-void PopulateBodySpaceTree(nc::asset::BodySpaceNode* bodySpaceNode, const aiNode* inputNode, nc::asset::BodySpaceNode* parent, uint32_t* size)
+void PopulateBodySpaceTree(nc::asset::BodySpaceNode* bodySpaceNode, const aiNode* inputNode, nc::asset::BodySpaceNode* parent, size_t* size)
 {
     if (!inputNode)
     {
